refactor: exctract getting colors to separated function
