+++ src/components/__tests__/ListItem.test.js
@@ -4,7 +4,7 @@ import renderer from 'react-test-renderer';
 import { Text, View } from 'react-native';
 import ListItem from '../List/ListItem.tsx';
 import ListIcon from '../List/ListIcon.tsx';
-import Chip from '../Chip';
+import Chip from '../Chip/Chip';
 import { red500 } from '../../styles/themes/v2/colors';
 
 const styles = StyleSheet.create({
